chore: Refactor and modularize server code
